+++ lib/crypto_vault.dart
@@ -1,14 +1,15 @@
 library crypto_vault;
 
-export 'data/dto/exception/crypto_vault_exception.dart';
-export 'data/dto/model/crypto_key.dart';
-export 'data/enum/rsa_digest.dart';
-export 'data/enum/rsa_encoding.dart';
-export 'data/repositories/crypto_aes_repository.dart';
-export 'data/repositories/crypto_aes_repository_impl.dart';
-export 'data/repositories/crypto_rsa_repository.dart';
-export 'data/repositories/crypto_rsa_repository_impl.dart';
-export 'data/repositories/crypto_ed25519_repository.dart';
-export 'data/repositories/crypto_ed25519_repository_impl.dart';
-
+export 'src/crypto_vault_aes.dart';
+export 'src/crypto_vault_ed25519.dart';
 export 'src/crypto_vault_ec.dart';
+export 'src/crypto_vault_rsa.dart';
+export 'src/data/repositories/crypto_vault_aes_default.dart';
+export 'src/data/repositories/crypto_vault_ed25519_default.dart';
+export 'src/data/repositories/crypto_vault_ec_default.dart';
+export 'src/data/repositories/crypto_vault_rsa_default.dart';
+export 'src/domain/crypto_vault_signature.dart';
+export 'src/domain/entities/crypto_key.dart';
+export 'src/domain/enums/crypto_vault_rsa_digest.dart';
+export 'src/domain/enums/crypto_vault_rsa_encoding.dart';
+export 'src/domain/exceptions/crypto_vault_exception.dart';