+++ lib/data/dto/exception/crypto_vault_exception.dart
@@ -1,9 +1,0 @@
-class CryptoVaultException implements Exception {
-  String code;
-  String message;
-
-  CryptoVaultException({
-    required this.code,
-    required this.message,
-  });
-}