+++ lib/data/enum/rsa_digest.dart
@@ -1,4 +1,0 @@
-enum CoreCryptoRSADigest {
-  sha1,
-  sha256,
-}