+++ lib/src/domain/entities/crypto_key.dart
@@ -2,15 +2,15 @@ class CryptoKey {
   final String publicKey;
   final String privateKey;
 
-  CryptoKey({
+  const CryptoKey({
     required this.publicKey,
     required this.privateKey,
   });
 
   Map<String, dynamic> toJson() {
     return {
-      "privateKey": privateKey,
-      "publicKey": publicKey,
+      'privateKey': privateKey,
+      'publicKey': publicKey,
     };
   }
 }