+++ lib/src/domain/enums/crypto_vault_rsa_encoding.dart
@@ -1,0 +1,4 @@
+enum CryptoVaultRsaEncoding {
+  pkcs1,
+  oaep,
+}